removes RegMenu.js
